Applcation for the Solid Project RDF Task Force
Hello! I am Denis from Russia.

I would like to join Solid Project RDF task force.

I have had some experience in the past with programming tests for CAD program in C++, web development with Java, PHP and JavaScript; maps and newspaper layout design.

Starting from 2016 I use Racket and Clojure for data processing, and starting from 2018 - elements of semantic web technology in my projects - RDF, OWL, SPARQL..

Speaking of interests, for several years, when relations between Russia and West were more or less positive, I have been engaged in information support of activities in the border area of Russia and Norway, with an aim to develop it, using strong sides of both worlds.

Later I switched to a system biology modelling, with an idea to develop method for describing cell and organism processes by the code, on the base of SBGN diagram languages and BioPAX ontology (this idea is still with me for the better times)

Since that I have interest in Semantic Web technologies and approach, using it in various hobby projects (like Telegram bot for city historical buildings over RDF triplestore), and business (analysis of posts in Vkontakte local advertisement group).

My dream is to have a "web of data", when individuals and cooperatives store their data in the pods and exchange/combine it in collaborative projects.

Solidifying the Solid Project side - is, for my opinion, the important starting step, to understand the model, how a topic data, describing several facets of reality can be gathered in its entirety in a pod and further used by applications. 

And as a person that still knows so little about Semantic Web and ways to work with it - I would have an excellent opportunity to learn these things by the work and from the colleagues.